+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 *   Add HTML bookmarks as an export format
 *   Rework widget item layout for consistent touch targets
+*   Add attribution for libraries used in the app
+*   Redesign settings screen
+*   Update to Compose January releases
 
 ## [1.38.0] - 2023-11-20